fix some bug in variable type

+++ Shaw/Assets/Character/Character.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Character : MonoBehaviour
 {
     public int Hp = 0, Mana = 0, Money = 0;
-    public double Speed = 0.0;
+    public float Speed = 0.0f;
     public List<GameObject> prefab = new List<GameObject>();
 
 }
